Separate remaining transcription test suites
Move live-client and SpeechAnalyzer unit tests out of production modules without changing their behavior.
